refactor(keyframes): extract DopesheetEditor state-cluster hooks
Pull four cohesive state clusters out of DopesheetEditor into
dedicated hooks: the local/global frame inputs (with their commit
planners and Enter/Escape handling), the graph view state (visible
curves with localStorage persistence, ruler unit, handle/zoom
toggles, curve toggle callbacks), property filters (group visibility,
keyframed-only mode, per-property locks with normalization effects),
and the group accordion expansion (auto-opening the active
property's group). Drops the orchestrator from 3,490 to 3,191 LOC and
exposes the state architecture at the top of the function as a
short, named chain of hooks. No behavior change.
